Implement a generic Levenshtein's distance algorithm
This implementation works with any type that implements the
`std::cmp::Eq` trait.

Though unoptimized, three implementations of the algorithm are provided:
1. Naive recursive implementation (complexity O(3^n))
2. Tabulation based dynamic programming implementation (complexity O(n^2))
2. Memoization based dynamic programming implementation (complexity O(n^2))

It is strongly recommended to avoid using the naive recursive
implementation because of it's horribly poor performance. In general,
the DP implementations fare better though they have a space complexity
of O(n^2).

The recommended implementation is the memoization one because in theory
it won't compute distances that it doesn't need to.